don't cache lfa
