+++ .eslintignore
@@ -1,3 +1,4 @@
+node_modules
 **/node_modules
 coverage
 packages/apps/react/dist
