+++ packages/baselines/function-approximators/generic/linear/index.spec.js
@@ -1,13 +1,42 @@
 const Linear = require('.');
-const Fourier = require('./bases/fourier');
+const Fourier = require('./bases/fourier-cos');
 
-test('performs update', () => {
-  const basis = new Fourier({
-    variables: 1,
-    order: 4,
-  });
-  const linear = new Linear({ alpha: 0.1, basis });
-  expect(linear.call([0])).toEqual(0);
-  linear.update([0], 1);
-  expect(linear.call([0])).toEqual(0.5);
+const basis = new Fourier({ variables: 2, order: 1 });
+
+test('call', () => {
+  const fa = new Linear({ alpha: 0.1, basis });
+  expect(fa.call([0, 1])).toEqual(0);
+  fa.setParameters([1, 1, 0, 0]);
+  expect(fa.call([0, 1])).toEqual(2);
+});
+
+test('update', () => {
+  const fa = new Linear({ alpha: 0.1, basis });
+  const args = [0, 1];
+  expect(fa.call(args)).toEqual(0);
+  fa.update(args, 10);
+  expect(fa.call(args)).toEqual(0.1 * 10 * basis.getNumberOfFeatures());
+});
+
+test('gradient', () => {
+  const fa = new Linear({ alpha: 0.1, basis });
+  expect(fa.gradient([0, 1])).toEqual([1, 1, -1, -1]);
+});
+
+test('getParameters', () => {
+  const fa = new Linear({ alpha: 0.1, basis });
+  expect(fa.getParameters()).toEqual([0, 0, 0, 0]);
+});
+
+test('setParameters', () => {
+  const fa = new Linear({ alpha: 0.1, basis });
+  fa.setParameters([1, 2, 3, 4]);
+  expect(fa.getParameters()).toEqual([1, 2, 3, 4]);
+});
+
+test('updatesParameters', () => {
+  const fa = new Linear({ alpha: 0.1, basis });
+  const args = [0, 1];
+  fa.updateParameters([1, 2, 3, 4]);
+  expect(fa.call(args)).toEqual(-0.4);
 });
