+++ tests/integration/.gitignore
@@ -1,0 +1,1 @@
+*results*.txt
